+++ test/ignitor.spec.js
@@ -30,8 +30,7 @@ test.group('Ignitor', (group) => {
     fold.ioc.fake('Adonis/Src/Exception', () => {
       return { bind () {} }
     })
-
-    // process.removeAllListeners('unhandledRejection')
+    process.removeAllListeners('unhandledRejection')
   })
 
   test('register app root', (assert) => {
@@ -91,7 +90,7 @@ test.group('Ignitor', (group) => {
     assert.plan(1)
     const ignitor = new Ignitor()
     try {
-      await ignitor.fireHttpServer()
+      await ignitor.fire()
     } catch ({ message }) {
       assert.equal(message, 'Cannot start http server, make sure to register the app root inside server.js file')
     }
@@ -300,59 +299,73 @@ test.group('Ignitor', (group) => {
   test('load ace commands when loadCommands method is called', async (assert) => {
     assert.plan(1)
     const ignitor = new Ignitor(fold)
-    ignitor.appRoot(path.join(__dirname, './'))
+    ignitor
+      .appRoot(path.join(__dirname, '../'))
+      .appFile('test/start/app.js')
+
     ignitor._preLoadFiles = []
-    const appFile = require(path.join(__dirname, ignitor._appFile))
-    appFile.aceProviders = ['Adonis/Src/Command']
-    try {
-      await ignitor.loadCommands().fire()
-    } catch ({ message }) {
-      assert.equal(message, `Cannot find module 'Adonis/Src/Command'`)
-    }
+
+    const appFile = require(path.join(__dirname, '../', ignitor._appFile))
+    appFile.commands = ['App/Commands/Greet']
+
+    fold.ioc.fake('App/Commands/Greet', function () {
+      assert.isTrue(true)
+
+      const { Command } = require('@adonisjs/ace')
+      return class FooCommand extends Command {
+        static get signature () {
+          return 'foo'
+        }
+      }
+    })
+
+    await ignitor.loadCommands().fire()
   })
 
-  test('setup exception handler to adonis provider when there is no exceptions handler', (assert) => {
+  test('setup exception handler to base exceptions handler when custom one doesnt exists', (assert) => {
     const ignitor = new Ignitor(fold)
     ignitor.appRoot(path.join(__dirname, './'))
-    assert.plan(2)
+    assert.plan(1)
 
-    class Exception {
-      bind (name, handler) {
-        assert.equal(name, '*')
-        assert.equal(handler, '@provider:Adonis/Exceptions/Handler')
+    class BaseHandler {
+      handle () {}
+      report () {}
+    }
+    fold.ioc.fake('Adonis/Exceptions/BaseExceptionHandler', () => BaseHandler)
+
+    class Server {
+      setExceptionHandler (handler) {
+        assert.deepEqual(handler, BaseHandler)
       }
     }
 
-    fold.ioc.fake('Adonis/Src/Exception', function () {
-      return new Exception()
-    })
-
     ignitor._setPackageFile()
     ignitor._setupResolver()
-    ignitor._setupExceptionsHandler()
+    ignitor._setupExceptionsHandler(new Server())
   })
 
   test('setup exception handler to global handler when defined', async (assert) => {
     const ignitor = new Ignitor(fold)
 
     ignitor.appRoot(path.join(__dirname, './'))
-    assert.plan(2)
-    await fs.outputFile(path.join(__dirname, './app/Exceptions/Handler.js'), 'module.exports = {}')
+    assert.plan(1)
 
-    class Exception {
-      bind (name, handler) {
-        assert.equal(name, '*')
-        assert.equal(handler, 'App/Exceptions/Handler')
+    await fs.outputFile(path.join(__dirname, './app/Exceptions/Handler.js'), `
+      module.exports = class CustomHandler {
+        handle () {}
+        report () {}
+      }
+    `)
+
+    class Server {
+      setExceptionHandler (handler) {
+        assert.equal(handler.name, 'CustomHandler')
       }
     }
 
-    fold.ioc.fake('Adonis/Src/Exception', function () {
-      return new Exception()
-    })
-
     ignitor._setPackageFile()
     ignitor._registerAutoloadedDirectories()
-    ignitor._setupExceptionsHandler()
+    ignitor._setupExceptionsHandler(new Server())
     await fs.remove(path.join(__dirname, './app'))
   })
 
@@ -377,10 +390,17 @@ test.group('Ignitor', (group) => {
       listen (h, p, cb) { cb() }
       getInstance () {
         return {
-          once () {}
+          once (event, cb) { cb() }
         }
       }
+      setExceptionHandler () {}
     }
+
+    class BaseHandler {
+      handle () {}
+      report () {}
+    }
+    fold.ioc.fake('Adonis/Exceptions/BaseExceptionHandler', () => BaseHandler)
 
     const events = []
     hooks.before.httpServer(() => {
@@ -415,7 +435,15 @@ test.group('Ignitor', (group) => {
       handle () {}
 
       listen (h, p, cb) { cb() }
+
+      setExceptionHandler () {}
     }
+
+    class BaseHandler {
+      handle () {}
+      report () {}
+    }
+    fold.ioc.fake('Adonis/Exceptions/BaseExceptionHandler', () => BaseHandler)
 
     fold.ioc.fake('Adonis/Src/Server', () => new Server())
     fold.ioc.fake('Adonis/Src/Env', () => new Env())
@@ -428,6 +456,38 @@ test.group('Ignitor', (group) => {
       return server
     })
 
+    server.close()
     assert.deepEqual(customInstance, server)
+  })
+
+  test('do not swallow errors in preloading files', (assert) => {
+    assert.plan(1)
+
+    const ignitor = new Ignitor(fold)
+    ignitor.appRoot(path.join(__dirname, './'))
+    ignitor.preLoad('start/emitsError.js')
+    ignitor._optionals.push('start/emitsError.js')
+
+    try {
+      ignitor._loadPreLoadFiles()
+    } catch (error) {
+      assert.equal(error.code, 'MODULE_NOT_FOUND')
+    }
+  })
+
+  test('load hooks file when it exists', async (assert) => {
+    assert.plan(1)
+    await fs.outputFile(path.join(__dirname, './start/hooks.js'), `
+      global.hooksLoaded = true
+    `)
+
+    const ignitor = new Ignitor(fold)
+    ignitor.appRoot(path.join(__dirname, './'))
+    ignitor._loadHooksFileIfAny()
+
+    assert.isTrue(global.hooksLoaded)
+
+    await fs.remove(path.join(__dirname, './start/hooks.js'))
+    delete global.hooksLoaded
   })
 })